Update strategy.js
I came across the need to request different scopes for certian types of users in my system.  Changing the name of the strategy allows me to easily handle this scenario.  

In my case, the first user in for a team sets up a project in my app... the rest just need to verify they have access to the same slack team.